Починить хеширование и хеш-таблицу в vol4
Модульная арифметика считалась в int при модуле Integer.MAX_VALUE, поэтому
произведение переполнялось уже на седьмой степени основания: таблица степеней
с этого места содержала мусор и отрицательные значения, а хеш длинного текста
получался отрицательным. Вычисления переведены в long. Разность префиксных
хешей могла быть отрицательной, и остаток от деления сохранял знак — результат
нормализован. Хеширование пустой строки падало с выходом за границу массива.

Наивный поиск подстроки не проверял последние позиции текста: search("abc","bc")
возвращал пусто. После частичного совпадения индекс прыгал на длину совпавшего
куска вместо одного шага, из-за чего search("aab","ab") тоже ничего не находил.

Хеш-таблица при повторной записи по существующему ключу молча теряла новое
значение и всё равно увеличивала size, поэтому размер считал вызовы put, а не
записи. Удаление отсутствующего ключа возвращало значение первого элемента
корзины — из-за этого частотный анализ «Войны и мира» систематически врал.
Удаление единственного элемента корзины не уменьшало size: break стоял до
декремента. Теперь топ-10 слов выглядит правдоподобно: сказал 411, князь 274,
Андрей 217, Пьер 188.

MainMap дважды открывал поток к файлу на 841 КБ и не закрывал его, а
InputStreamReader создавался без кодировки и на не-UTF-8 платформе читал
русский текст неверно.

Ошибки в PolynomialSearchCached и отсутствие перехеширования оставлены как
задания и закреплены @Disabled-тестами. В модуле было ноль тестов, стало 21.

+++ vol4/src/main/java/ru/mifi/practice/vol4/hash/Hash.java
@@ -36,14 +36,14 @@ public interface Hash {
 
         @Override
         public int hash(String text, Counter counter) {
-            int hash = 0;
-            int next = 1;
+            long hash = 0;
+            long next = 1;
             for (int i = 0; i < text.length(); i++) {
                 hash = (hash + next * text.charAt(i)) % mod;
                 next = next * k % mod;
                 counter.increment();
             }
-            return hash;
+            return (int) hash;
         }
     }
